+++ CHANGES.md
@@ -1,4 +1,7 @@
 
+  - [Search]: solve one-player games using graph traversal algorithms
+    (depth-first search, breadth-first search, iterative deepening
+     depth-first search)
   - [Traverse.Bfs]: new function `{fold,iter}_component_dist` to
     perform a breadth-first traversal with the distance from the source